Move left networking e2e tests to test/e2e/network
