+++ test/e2e/BUILD
@@ -59,18 +59,14 @@ go_library(
         "generated_clientset.go",
         "gke_local_ssd.go",
         "gke_node_pools.go",
-        "ingress.go",
         "limit_range.go",
         "network_partition.go",
-        "no-snat.go",
         "pod_gc.go",
         "podpreset.go",
         "pods.go",
         "pre_stop.go",
         "resource_quota.go",
         "service_accounts.go",
-        "service_latency.go",
-        "serviceloadbalancers.go",
         "ssh.go",
         "ubernetes_lite.go",
     ],
@@ -122,10 +118,8 @@ go_library(
         "//vendor/k8s.io/apiserver/pkg/authentication/serviceaccount:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes/typed/certificates/v1beta1:go_default_library",
-        "//vendor/k8s.io/client-go/rest:go_default_library",
         "//vendor/k8s.io/client-go/tools/cache:go_default_library",
         "//vendor/k8s.io/client-go/util/cert:go_default_library",
-        "//vendor/k8s.io/client-go/util/flowcontrol:go_default_library",
     ],
 )
